[Chore]#22 : rename UserEntity, UserInfoEntity -> UserDto, UserDtoEntity
