+++ client/app/src/main/java/com/example/healthc/data/dto/auth/UserDto.kt
@@ -1,6 +1,6 @@
 package com.example.healthc.data.dto.auth
 
-data class UserEntity (
+data class UserDto (
     val email : String,
     val password: String
 )
