+++ client/app/src/main/java/com/example/healthc/data/local/UserInfoDataBase.kt
@@ -3,10 +3,10 @@ package com.example.healthc.data.local
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.example.healthc.data.local.entity.LocalUserInfoEntity
+import com.example.healthc.data.local.entity.UserInfoEntity
 
 @Database(
-    entities = [LocalUserInfoEntity::class],
+    entities = [UserInfoEntity::class],
     version = 1
 )
 @TypeConverters(Converters::class)
