+++ client/app/src/main/java/com/example/healthc/data/source/auth/SignInDataSource.kt
@@ -1,9 +1,9 @@
 package com.example.healthc.data.source.auth
 
-import com.example.healthc.data.dto.auth.UserEntity
+import com.example.healthc.data.dto.auth.UserDto
 import com.example.healthc.domain.utils.Resource
 import com.google.firebase.auth.FirebaseUser
 
 interface SignInDataSource {
-    suspend fun signIn(userEntity: UserEntity) : Resource<FirebaseUser>
+    suspend fun signIn(userDto: UserDto) : Resource<FirebaseUser>
 }
