+++ client/app/src/main/java/com/example/healthc/data/source/auth/SignInDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.example.healthc.data.source.auth
 
-import com.example.healthc.data.dto.auth.UserEntity
+import com.example.healthc.data.dto.auth.UserDto
 import com.example.healthc.domain.utils.Resource
 import com.example.healthc.data.utils.await
 import com.google.firebase.auth.FirebaseAuth
@@ -10,10 +10,10 @@ import javax.inject.Inject
 class SignInDataSourceImpl @Inject constructor(
     private val firebaseAuth: FirebaseAuth
 ): SignInDataSource {
-    override suspend fun signIn(userEntity: UserEntity): Resource<FirebaseUser> {
+    override suspend fun signIn(userDto: UserDto): Resource<FirebaseUser> {
         return try{
             val result = firebaseAuth.signInWithEmailAndPassword(
-                userEntity.email, userEntity.password
+                userDto.email, userDto.password
             ).await()
             Resource.Success(checkNotNull(result.user))
         }
