+++ client/app/src/main/java/com/example/healthc/data/source/auth/SignUpDataSource.kt
@@ -1,10 +1,10 @@
 package com.example.healthc.data.source.auth
 
-import com.example.healthc.data.dto.auth.UserEntity
-import com.example.healthc.data.dto.auth.UserInfoEntity
+import com.example.healthc.data.dto.auth.UserDto
+import com.example.healthc.data.dto.auth.UserInfoDto
 import com.example.healthc.domain.utils.Resource
 import com.google.firebase.auth.FirebaseUser
 
 interface SignUpDataSource {
-    suspend fun signUp(userEntity: UserEntity, userInfoEntity: UserInfoEntity) : Resource<FirebaseUser>
+    suspend fun signUp(userDto: UserDto, userInfoDto: UserInfoDto) : Resource<FirebaseUser>
 }
